Removed qrequestfiles in an attempt to solve SP issues
